Temporarily modified Bullet class's Draw method to recieve Double parameters. Moved some code from MainPage constructor to OnNavigatedTo method to accomodate for information passing. Modified SetImage to set proper values.

+++ AlienInvaders/Bullet.cs
@@ -59,7 +59,7 @@ namespace AlienInvaders
             return false;
         }
 
-        public bool Draw(float xPosition, float yPosition)
+        public bool Draw(double xPosition, double yPosition)
         {
             if(IsAlive == false)
             {
